requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.4
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
jdcal==1.4.1
mysqlclient==1.4.2
odfpy==1.4.0
openpyxl==2.6.3
Pillow==6.1.0
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
